requirements.txt
pymongo==4.7.0
passlib==1.7.4
python-dotenv==1.0.1
uvicorn==0.29.0
slowapi==0.1.9
pydantic==2.7.1
pytz==2024.1
pymsteams==0.2.2
python-multipart
fastapi==0.109.0
python-jose
pydantic[email]
